move KeySelectionValueSelectionValueAttribute to Manager
